+++ source/IO/StreamArchive.cpp
@@ -40,7 +40,7 @@ namespace rePlayer
         return nullptr;
     }
 
-    size_t StreamArchive::Read(void* buffer, size_t size)
+    uint64_t StreamArchive::Read(void* buffer, uint64_t size)
     {
         if (m_streamMemory.IsValid())
             return m_streamMemory->Read(buffer, size);
@@ -63,7 +63,7 @@ namespace rePlayer
                 m_entryDataBlockPosition = 0;
             }
             auto sizeToCopy = Min(sizeLeft, remainingSize);
-            memcpy(dest, m_entryDataBlock + m_entryDataBlockPosition, sizeToCopy);
+            memcpy(dest, m_entryDataBlock + m_entryDataBlockPosition, size_t(sizeToCopy));
             dest += sizeToCopy;
             m_entryPosition += sizeToCopy;
             m_entryDataBlockPosition += sizeToCopy;
@@ -120,12 +120,12 @@ namespace rePlayer
         return Status::kOk;
     }
 
-    size_t StreamArchive::GetSize() const
+    uint64_t StreamArchive::GetSize() const
     {
         return m_entrySize;
     }
 
-    size_t StreamArchive::GetPosition() const
+    uint64_t StreamArchive::GetPosition() const
     {
         if (m_streamMemory.IsValid())
             return m_streamMemory->GetPosition();
