+++ source/IO/StreamArchive.h
@@ -16,11 +16,11 @@ namespace rePlayer
         static [[nodiscard]] SmartPtr<StreamArchive> Create(const std::string& filename, bool isPackage);
         static [[nodiscard]] SmartPtr<StreamArchive> Create(io::Stream* stream, bool isPackage);
 
-        size_t Read(void* buffer, size_t size) final;
+        uint64_t Read(void* buffer, uint64_t size) final;
         Status Seek(int64_t offset, SeekWhence whence) final;
 
-        [[nodiscard]] size_t GetSize() const final;
-        [[nodiscard]] size_t GetPosition() const final;
+        [[nodiscard]] uint64_t GetSize() const final;
+        [[nodiscard]] uint64_t GetPosition() const final;
 
         [[nodiscard]] const std::string& GetName() const final { return m_entryFilename; }
 
@@ -57,7 +57,7 @@ namespace rePlayer
         std::string m_entryFilename;
         bool m_isPackage = false;
         uint32_t m_entryIndex = 0;
-        size_t m_entrySize;
+        uint64_t m_entrySize;
         int64_t m_entryPosition = 0;
         uint8_t* m_entryDataBlock;
         size_t m_entryDataBlockSize = 0;
